Refactor posizione package
