+++ app/src/main/java/it/unimib/brewday/ui/gestisciAttrezzi/GestisciAttrezziViewModelFactory.java
@@ -1,10 +1,12 @@
-package it.unimib.brewday.ui;
+package it.unimib.brewday.ui.gestisciAttrezzi;
 
 import android.content.Context;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
+
+import it.unimib.brewday.ui.gestisciAttrezzi.GestisciAttrezziViewModel;
 
 public class GestisciAttrezziViewModelFactory implements ViewModelProvider.Factory {
