+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
         android:id="@+id/fragment_container_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:name="it.unimib.brewday.ui.GestisciAttrezziFragment" />
+        android:name="it.unimib.brewday.ui.gestisciAttrezzi.GestisciAttrezziFragment" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
