+++ app/src/main/res/layout/fragment_gestisci_attrezzi.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.GestisciAttrezziFragment">
+    tools:context=".ui.gestisciAttrezzi.GestisciAttrezziFragment">
 
     <ImageButton
         android:id="@+id/gestisciAttrezziFragment_imageButton_add"
